Major clean up for the Python project.
This commit moves a bunch of files around to try to fit more closely
with the Python packaging guidelines. It also adds a useful README.md, a
LICENSE file containing the project's MIT license, and generally tries to
restructure the code to be clearer and more Pythonic.

There is still work to be done so that the project can be submitted to Pypi and
installed locally. Also, some of the tests are still commented out. This is
definitely a work in progress.